Update FERC 1 and 714 to pull 2020 data
The FERC 1 spider has an explicit list of years it tries to scrape.
Maybe this isn't how it should really work... but I updated it to
include 2020 data.

For the FERC 714 data, FERC seems to choose a new hand-crafted URL every
year, which isn't easily scraped from their web page. So I updated the
hard coded URL. Not great.