Pull out async retries into their own util function.
In theory we *should* be using `async with` for the ClientSession
stuff because the various requests are async context managers. But
they work with plain old `await`, and the docs claim we don't need
to call `release()` manually.

I didn't want to pull in aiohttp-retry because it seems like pulling
in a whole library just for retry logic is silly.

But at the same time the retries are kind of clunky.