Standardize filenames to be similar to other datasets
All of the other archivers use a file naming pattern like:

dataset-part1-part2.ext

where:
- dataset is the name of the dataset (e.g. ferc1, eia923, mshamines)
- part1 is the value of the first partition dimension contained in
  the file (e.g. ca, ny, wy, for states, 2019 if it's a year)
- part2 is the value of the second partition dimension contained in
  the file, if applicable.
- ext is the file name extension indicating file type.

I've changed the naming for phmsagas to follow the `dataset-*` part of
the naming convention, but have left the filename in place rather than
using only the dataset and start year, since if we just use the start
year it looks like years are missing in the archive, as each file
contains several years of data.

The dataset and start_year partitions are still used in datapackage.json
to refer to the individual resources for programmatic purposes.

This came up when working on #79