Update datapackage.json, but only sometimes (#205)
If we update `datapackage.json` to use the "canonical" URL of `/records/<record_id>/files/<file_name>` vs. whatever the hell `self.links.download` gives us (`/api/records/<record_id>(/draft)?/files/<file_name>(/content)?`), then we *do* want that to register as a change in the datapackage. But if we see the format stay the same & the record ID increment, then we don't want that to register as a change.