Don't do a weird song-and-dance to guarantee a fresh draft.
Using the new API for making a new draft avoids the 'there is already a draft' error, but leaves us with an empty draft to start which changes the semantics of the 'changes' list. Since we are aiming to handle 'starting draft in any state' we shouldn't rely on the old 'changes' semantics to decide when to publish.